requirements.txt
FFlask==2.0.1
SQLAlchemy==1.4.23
flask-sqlalchemy==2.5.1
python-dotenv==0.19.0
flask-migrate==3.1.0
flask-cors==3.0.10
